requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
langchain
langchain-google-genai
